+++ Samples~/ImportTester/Code/InputListener.cs
@@ -5,8 +5,10 @@ namespace At.Ac.FhStp.Import3D.ImportTester
 {
     public class InputListener : MonoBehaviour
     {
-        [SerializeField] private UnityEvent<LinearDirection> modelSwitchInput;
-
+        [SerializeField] private UnityEvent<LinearDirection> modelSwitchInput =
+                new UnityEvent<LinearDirection>();
+        [SerializeField] private UnityEvent reloadModelsInput =
+                new UnityEvent();
 
         private void Update()
         {
@@ -14,6 +16,8 @@ namespace At.Ac.FhStp.Import3D.ImportTester
                 modelSwitchInput.Invoke(LinearDirection.Previous);
             else if (Input.GetKeyDown(KeyCode.E))
                 modelSwitchInput.Invoke(LinearDirection.Next);
+            else if(Input.GetKeyDown(KeyCode.R))
+                reloadModelsInput.Invoke();
         }
     }
 }